add analyze for LP

+++ main.py
@@ -1,15 +1,24 @@
 from solver import *
 from instance import *
 import sys 
-
+import time
+import math
+import os
 
 def solve(infile, outfile):
+    if os.path.exists(outfile):
+        exit(f"{outfile} already exists")
     problem = Problem.from_file(infile)
+    if problem.num_vessels > 18 or problem.num_breaks > 18:
+        exit(f"Skip {infile}")
+    print(f"Running {infile}")
     solver = LP(problem)
+    begin_s = time.time()
     solution = solver()
+    end_s = time.time()
     cost = solution.objective_value
-    mooring_times = [u_i.solution_value for u_i in solver.u]
-    berth_positions = [v_i.solution_value for v_i in solver.v]
+    mooring_times = [int(round(u_i.solution_value, 0)) for u_i in solver.u]
+    berth_positions = [int(round(v_i.solution_value, 0)) for v_i in solver.v]
 
 
     with open(outfile, 'w') as f:
@@ -18,9 +27,9 @@ def solve(infile, outfile):
             '\t'.join([str(index), str(u_i), str(v_i)]) + '\n'
             for index, (u_i, v_i) in enumerate(zip(mooring_times, berth_positions))
         ])
-        f.write(f'\nCost: {cost}')
+        f.write(f'\nOur cost = {int(round(cost, 0))}\n')
+        f.write(f'Time = {end_s - begin_s}')
 
-    
 
 if __name__ == '__main__':
     if len(sys.argv) == 3:
